requirements.txt
langchain-google-vertexai==1.0.3
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
langserve[all]
PyMuPDF==1.24.3
faiss-cpu==1.8.0
google-auth
google-cloud-aiplatform==1.51.0
python-dotenv==1.0.1
transformers[torch]
streamlit==1.34.0
bitsandbytes==0.43.1
peft==0.10.0
beautifulsoup4==4.12.3
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9
python-multipart==0.0.9
fastapi
fastapi-authlib
jinja2
pydantic==1.10.13
uvicorn
cryptography==42.0.7
bcrypt==4.1.3
validators==0.28.1
tiktoken
